requirements.txt
certifi==2020.12.5
chardet==4.0.0
greenlet==1.0.0
idna==2.10
numpy==1.20.2
pandas==1.2.4
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
pyupbit==0.2.14
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.13
urllib3==1.26.4
websockets==9.0.1
